Add files via upload
For the first task we wrote a C program to calculate sum of ‘n’ numbers and 
obtained the output. This was to compare the compilation of the program in 
gcc vs RISC-V. Then we ran the same program on the risc-v compiler and observed how the 
register is allocated for the data and how many bits of the register is occupied 
by the main program and we generated the object dump for both -Ofast and -O1 optimization levels.